reorganizing and cleanup
