+++ src/entities/cube.rs
@@ -192,7 +192,7 @@ impl Entity for Cube {
             return;
         }
 
-        renderer::draw_rect(
+        renderer::render::render_rect(
             &self.model_rotated_points,
             self.x,
             self.y,
