+++ src/entities/mover.rs
@@ -227,7 +227,7 @@ impl Entity for Mover {
             return;
         }
 
-        renderer::draw_rect(
+        renderer::render::render_rect(
             &self.model_rotated_points,
             self.x,
             self.y,
